feat: add logout button
