+++ pages/login.js
@@ -1,12 +1,14 @@
 import React, { useState } from "react";
+
+import { signIn } from "../utils/auth";
+import { useRouter } from "next/router";
+
+// MUI
 import Box from "@mui/material/Box";
 import TextField from "@mui/material/TextField";
 import Button from "@mui/material/Button";
 import Alert from "@mui/material/Alert";
 import Stack from "@mui/material/Stack";
-
-import { signIn } from "../utils/auth";
-import { useRouter } from "next/router";
 
 function logIn() {
   const [email, setEmail] = useState("");
