+++ pages/roadmap.js
@@ -1,0 +1,5 @@
+function Roadmap() {
+  return <h1>Soy el terribel Roadmap</h1>;
+}
+
+export default Roadmap;
