--- pages/roadmaps.js
@@ -1,3 +1,0 @@
-function Roadmap() {
-	return <h1>Soy el terribel Roadmap</h1>;
-}
